UserProfile component was split by 2: UserProfile and UserPasswordProfile, both are included in UserProfilePage
