+++ src/actions/profileEditAction.js
@@ -1,4 +1,4 @@
-  import { profileEditService } from 'services';
+import { profileEditService } from 'services';
 import { DIFFERENT_OLD_PASSWORD } from 'services/profileEditService';
 import { SubmissionError, clearFields } from 'redux-form'
 
@@ -7,6 +7,10 @@ import { updateUser } from './sessionAction';
 export const PROFILE_EDIT_REQUEST = 'PROFILE_EDIT_REQUEST'
 export const PROFILE_EDIT_SUCCESS = 'PROFILE_EDIT_SUCCESS'
 export const PROFILE_EDIT_FAILURE = 'PROFILE_EDIT_FAILURE'
+
+export const PROFILE_PASSWORD_EDIT_REQUEST = 'PROFILE_PASSWORD_EDIT_REQUEST'
+export const PROFILE_PASSWORD_EDIT_SUCCESS = 'PROFILE_PASSWORD_EDIT_SUCCESS'
+export const PROFILE_PASSWORD_EDIT_FAILURE = 'PROFILE_PASSWORD_EDIT_FAILURE'
 
 //Edit user's fields
 export const profileEdit = (profile) => {
@@ -23,12 +27,6 @@ export const profileEdit = (profile) => {
         },
         error => {
           dispatch(failure(error))
-          if (error === DIFFERENT_OLD_PASSWORD){
-            throw new SubmissionError({
-              old_password: 'Senha antiga não confere',
-              _error: 'Senha antiga não confere'
-            })
-          }
           throw new SubmissionError({
             _error: 'Não existe conta associada com este email'
           })
@@ -44,7 +42,7 @@ export const profileEdit = (profile) => {
 //Redefine password_confirmation
 export const redefineUserPassword = (profile) => {
   return dispatch => {
-    dispatch(requestEditProfile(profile))
+    dispatch(requestRedefineUserPassword(profile))
     return profileEditService.profileEdit(profile)
       .then(
         data => {
@@ -69,7 +67,7 @@ export const redefineUserPassword = (profile) => {
       )
   }
 
-  function requestEditProfile(){ return { type: PROFILE_EDIT_REQUEST } }
+  function requestRedefineUserPassword(){ return { type: PROFILE_EDIT_REQUEST } }
   function success(){ return { type: PROFILE_EDIT_SUCCESS } }
   function failure(error){ return { type: PROFILE_EDIT_FAILURE, error } }
 }
