+++ src/components/userprofile/UserProfile.js
@@ -28,8 +28,6 @@ const UserProfile = props => {
   const { handleSubmit, user, submitSucceeded } = props
 
     return (
-      <div className="row justify-content-center">
-        <div className="col-sm-8 col-md-8 col-lg-12 col-xs-12">
           <Form onSubmit={ handleSubmit }>
             <Container>
               <Row className="section-user-title">
@@ -87,89 +85,10 @@ const UserProfile = props => {
               </Row>
             </Container>
           </Form>
-
-          <Form>
-          <Container>
-              <Row  className="section-user-title">
-                <h4>Minha Conta</h4>
-              </Row>
-              <Row>
-                <Col sm="4">
-                  Email
-                </Col>
-                <Col sm="4">
-                  <Label>{user.email}</Label>
-                </Col>
-              </Row>
-              <Row className="sub-section-user-title">
-                <h5>Trocar senha</h5>
-              </Row>
-              <Row>
-                <Col sm="4" xs="12">
-                  <Field
-                    component={renderField}
-                    type="password"
-                    name="old_password"
-                    id="old_password"
-                    placeholder="Senha Atual"
-                    className="form-control"
-                  />
-                </Col>
-                <Col sm="4">
-                  <Field
-                    component={renderField}
-                    type="password"
-                    name="new_password"
-                    id="new_password"
-                    placeholder="Nova senha"
-                    className="form-control"
-                  />
-                </Col>
-                <Col sm="4">
-                  <Field
-                    component={renderField}
-                    type="password"
-                    name="password_confirmation"
-                    id="password_confirmation"
-                    placeholder="Confirme nova senha"
-                    className="form-control"
-                  />
-                </Col>
-              </Row>
-              <Row className="section-user-title">
-                <Col className="text-center">
-                  <Button>Redefinir senha</Button>
-                </Col>
-              </Row>
-              <div>
-                  { submitSucceeded && <UncontrolledAlert color="success">
-                    Usuário alterado com sucesso
-                    </UncontrolledAlert> }
-              </div>
-            </Container>
-          </Form>
-        </div>
-    </div>
   )
 }
 
-const validate = values => {
-  const errors = {}
 
-  if (values.new_password) {
-    if (values.new_password.length < 8){
-      errors.new_password = 'A nova senha deve conter no mínimo 8 dígitos'
-    } else if (!isNaN(values.new_password)){
-      errors.new_password = 'A nova senha não deve conter apenas números'
-    }
-  }
-
-  if (values.new_password && values.password_confirmation && values.new_password !== values.password_confirmation){
-    errors.password_confirmation = 'Senha e confirmação não coincidem'
-  }
-
-  return errors
-}
 
 const mapStateToProps = state => {
   const { user } = state.session.session
@@ -184,7 +103,6 @@ const mapStateToProps = state => {
 export default connect(
   mapStateToProps
 )
-  (reduxForm({
-  form : "profile",
-  validate
+(reduxForm({
+  form: 'profile'
 })(UserProfile));
